Logout instead of reloading when authentication expires
