+++ frontend/public/locales/en/translation.json
@@ -404,8 +404,8 @@
     "clusterPool.creation.ocp.size": "Cluster pool size",
     "clusterPool.creation.success.message": "{{name}} was successfully created.",
     "clusterPool.creation.success.title": "Cluster pool created",
-    "clusterPool.creation.validation.unique.name": "This cluster pool name must be unique in this namespace: {{0}}.",
     "clusterPool.creation.validation.runningCount.lessThanOrEqualSize": "The cluster pool running count must be less than or equal to the size.",
+    "clusterPool.creation.validation.unique.name": "This cluster pool name must be unique in this namespace: {{0}}.",
     "clusterPool.destroy": "Destroy cluster pool",
     "clusterPool.modal.scale.message": "Specify the desired size of <bold>{{clusterPoolName}}</bold>. Adjusting the size of the cluster pool will result in the creation or destruction of clusters; only unclaimed clusters will be destroyed if downsizing.",
     "clusterPool.modal.scale.poolName": "Cluster pool name",
