+++ frontend/src/atoms.tsx
@@ -3,6 +3,7 @@ import { AcmRoute } from '@stolostron/ui-components'
 import { CIM } from 'openshift-assisted-ui-lib'
 import { Fragment, ReactNode, useEffect, useMemo, useState } from 'react'
 import { atom, SetterOrUpdater, useRecoilState } from 'recoil'
+import { noop } from 'lodash'
 import { LoadingPage } from './components/LoadingPage'
 import {
     AgentClusterInstallApiVersion,
@@ -63,6 +64,7 @@ import {
     DiscoveryConfig,
     DiscoveryConfigApiVersion,
     DiscoveryConfigKind,
+    fetchGet,
     getBackendUrl,
     GitOpsCluster,
     GitOpsClusterApiVersion,
@@ -497,6 +499,13 @@ export function LoadData(props: { children?: ReactNode }) {
     }, [setSettings, setters])
 
     useEffect(() => {
+        function tokenExpired() {
+            if (process.env.NODE_ENV === 'production') {
+                logout()
+            } else {
+                window.location.href = `${getBackendUrl()}/login`
+            }
+        }
         function checkLoggedIn() {
             fetch(`${getBackendUrl()}/authenticated`, {
                 credentials: 'include',
@@ -507,20 +516,12 @@ export function LoadData(props: { children?: ReactNode }) {
                         case 200:
                             break
                         default:
-                            if (process.env.NODE_ENV === 'production') {
-                                window.location.reload()
-                            } else {
-                                window.location.href = `${getBackendUrl()}/login`
-                            }
+                            tokenExpired()
                             break
                     }
                 })
                 .catch(() => {
-                    if (process.env.NODE_ENV === 'production') {
-                        window.location.reload()
-                    } else {
-                        window.location.href = `${getBackendUrl()}/login`
-                    }
+                    tokenExpired()
                 })
                 .finally(() => {
                     setTimeout(checkLoggedIn, 30 * 1000)
@@ -542,4 +543,27 @@ export function usePolicies() {
         () => policies.filter((policy) => !policy.metadata.labels?.['policy.open-cluster-management.io/root-policy']),
         [policies]
     )
+}
+
+export async function logout() {
+    const tokenEndpointResult = await fetchGet<{ token_endpoint: string }>(getBackendUrl() + '/configure')
+    await fetchGet(getBackendUrl() + '/logout').catch(noop)
+
+    const iframe = document.createElement('iframe')
+    iframe.setAttribute('type', 'hidden')
+    iframe.name = 'hidden-form'
+    document.body.appendChild(iframe)
+
+    const form = document.createElement('form')
+    form.method = 'POST'
+    form.target = 'hidden-form'
+    const url = new URL(tokenEndpointResult.data.token_endpoint)
+    form.action = `${url.protocol}//${url.host}/logout`
+    document.body.appendChild(form)
+
+    form.submit()
+
+    await new Promise((resolve) => setTimeout(resolve, 500))
+
+    location.pathname = '/'
 }
